+++ packages/sheets/src/commands/commands/__tests__/set-style.command.spec.ts
@@ -32,7 +32,7 @@ import {
 } from '@univerjs/core';
 
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { SheetsSelectionsService } from '../../../services/selections/selection-manager.service';
+import { SheetsSelectionsService } from '../../../services/selections/selection.service';
 import { InsertSheetMutation } from '../../mutations/insert-sheet.mutation';
 import { SetRangeValuesMutation } from '../../mutations/set-range-values.mutation';
 import { InsertSheetCommand } from '../insert-sheet.command';
