+++ packages/sheets/src/commands/commands/insert-row-col.command.ts
@@ -32,7 +32,7 @@ import {
     RANGE_TYPE,
     sequenceExecute,
 } from '@univerjs/core';
-import { SheetsSelectionsService } from '../../services/selections/selection-manager.service';
+import { SheetsSelectionsService } from '../../services/selections/selection.service';
 import { SheetInterceptorService } from '../../services/sheet-interceptor/sheet-interceptor.service';
 import {
     InsertColMutation,
