+++ packages/sheets/src/commands/commands/set-worksheet-col-width.command.ts
@@ -26,7 +26,7 @@ import {
     Rectangle,
     sequenceExecute,
 } from '@univerjs/core';
-import { SheetsSelectionsService } from '../../services/selections/selection-manager.service';
+import { SheetsSelectionsService } from '../../services/selections/selection.service';
 import { SheetInterceptorService } from '../../services/sheet-interceptor/sheet-interceptor.service';
 import {
     SetWorksheetColWidthMutation,
